requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==11.1.0
python-dotenv==1.0.1
openpyxl==3.1.5
